исправление ошибок ребейза и под новую структуру
