+++ src/main/java/com/github/_1c_syntax/mdclasses/unmarshal/wrapper/DesignerCopyright.java
@@ -19,18 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with MDClasses.
  */
-package com.github._1c_syntax.mdclasses.mdo.wrapper;
+package com.github._1c_syntax.mdclasses.unmarshal.wrapper;
 
-import com.thoughtworks.xstream.annotations.XStreamAlias;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@XStreamAlias("item")
-public class DesignerContentItem {
-    @XStreamAlias("lang")
-    private String language = "";
-    @XStreamAlias("content")
-    private String content = "";
+public class DesignerCopyright extends DesignerContentItem {
 }
