+++ src/test/java/com/github/_1c_syntax/mdclasses/mdo/MDCommonTemplateTest.java
@@ -65,7 +65,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtGraphicalScheme() {
     var mdo = getMDObjectEDT("CommonTemplates/ГрафическаяСхема/ГрафическаяСхема.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ГрафическаяСхема",
-            "4333f027-4fc2-40a0-ae7d-48fbf0cea50b");
+      "4333f027-4fc2-40a0-ae7d-48fbf0cea50b");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.GRAPHICAL_SCHEME);
@@ -74,7 +74,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtGeoScheme() {
     var mdo = getMDObjectEDT("CommonTemplates/ГеографическаяСхема/ГеографическаяСхема.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ГеографическаяСхема",
-            "1d8d8dfc-e7c5-445a-a88d-6743faad2ab6");
+      "1d8d8dfc-e7c5-445a-a88d-6743faad2ab6");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.GEOGRAPHICAL_SCHEMA);
@@ -83,7 +83,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtAddin() {
     var mdo = getMDObjectEDT("CommonTemplates/ВнешняяКомпонента/ВнешняяКомпонента.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ВнешняяКомпонента",
-            "4a0dab22-affd-4887-a9b6-57a1e88f8377");
+      "4a0dab22-affd-4887-a9b6-57a1e88f8377");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.ADD_IN);
@@ -92,7 +92,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtTextDocument() {
     var mdo = getMDObjectEDT("CommonTemplates/ТекстовыйДокумент/ТекстовыйДокумент.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ТекстовыйДокумент",
-            "3084f392-8f90-4134-a82e-790e225aab29");
+      "3084f392-8f90-4134-a82e-790e225aab29");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.TEXT_DOCUMENT);
@@ -101,7 +101,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtDataCompositionScheme() {
     var mdo = getMDObjectEDT("CommonTemplates/СКД/СКД.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "СКД",
-            "8ae95178-7f50-4a77-aaf8-f8ffb72c65d4");
+      "8ae95178-7f50-4a77-aaf8-f8ffb72c65d4");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.DATA_COMPOSITION_SCHEME);
@@ -110,7 +110,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtDataCompositionAppearance() {
     var mdo = getMDObjectEDT("CommonTemplates/МакетОформления/МакетОформления.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "МакетОформления",
-            "f6bbaf46-bc77-412b-9440-3032bfc06c57");
+      "f6bbaf46-bc77-412b-9440-3032bfc06c57");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.DATA_COMPOSITION_APPEARANCE_TEMPLATE);
@@ -119,7 +119,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtBinaryData() {
     var mdo = getMDObjectEDT("CommonTemplates/ДвоичныеДанные/ДвоичныеДанные.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ДвоичныеДанные",
-            "f4ab9283-1110-4808-9cbf-40c71ebb88a2");
+      "f4ab9283-1110-4808-9cbf-40c71ebb88a2");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.BINARY_DATA);
@@ -128,7 +128,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtHtmlDocument() {
     var mdo = getMDObjectEDT("CommonTemplates/HTML/HTML.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "HTML",
-            "5d76084a-68fa-4579-91da-17d7ffab6225");
+      "5d76084a-68fa-4579-91da-17d7ffab6225");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.HTML_DOCUMENT);
@@ -137,7 +137,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtSpreadSheetDocument() {
     var mdo = getMDObjectEDT("CommonTemplates/ТабличныйДокумент/ТабличныйДокумент.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "ТабличныйДокумент",
-            "5b54ba38-ec04-4fc6-897f-48d36d0312a6");
+      "5b54ba38-ec04-4fc6-897f-48d36d0312a6");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.SPREADSHEET_DOCUMENT);
@@ -146,7 +146,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkEdtActiveDocument() {
     var mdo = getMDObjectEDT("CommonTemplates/Active/Active.mdo");
     checkBaseField(mdo, MDCommonTemplate.class, "Active",
-            "557665fc-86f5-44e1-9801-490cea841718");
+      "557665fc-86f5-44e1-9801-490cea841718");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.ACTIVE_DOCUMENT);
@@ -155,7 +155,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalGraphicalScheme() {
     var mdo = getMDObjectDesigner("CommonTemplates/ГрафическаяСхема.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "ГрафическаяСхема",
-            "4333f027-4fc2-40a0-ae7d-48fbf0cea50b");
+      "4333f027-4fc2-40a0-ae7d-48fbf0cea50b");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.GRAPHICAL_SCHEME);
@@ -164,7 +164,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalGeoScheme() {
     var mdo = getMDObjectDesigner("CommonTemplates/ГеографическаяСхема.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "ГеографическаяСхема",
-            "1d8d8dfc-e7c5-445a-a88d-6743faad2ab6");
+      "1d8d8dfc-e7c5-445a-a88d-6743faad2ab6");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.GEOGRAPHICAL_SCHEMA);
@@ -173,7 +173,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalAddin() {
     var mdo = getMDObjectDesigner("CommonTemplates/ВнешняяКомпонента.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "ВнешняяКомпонента",
-            "4a0dab22-affd-4887-a9b6-57a1e88f8377");
+      "4a0dab22-affd-4887-a9b6-57a1e88f8377");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.ADD_IN);
@@ -182,7 +182,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalTextDocument() {
     var mdo = getMDObjectDesigner("CommonTemplates/ТекстовыйДокумент.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "ТекстовыйДокумент",
-            "3084f392-8f90-4134-a82e-790e225aab29");
+      "3084f392-8f90-4134-a82e-790e225aab29");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.TEXT_DOCUMENT);
@@ -191,7 +191,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalDataCompositionScheme() {
     var mdo = getMDObjectDesigner("CommonTemplates/СКД.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "СКД",
-            "8ae95178-7f50-4a77-aaf8-f8ffb72c65d4");
+      "8ae95178-7f50-4a77-aaf8-f8ffb72c65d4");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.DATA_COMPOSITION_SCHEME);
@@ -200,7 +200,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalDataCompositionAppearance() {
     var mdo = getMDObjectDesigner("CommonTemplates/МакетОформления.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "МакетОформления",
-            "f6bbaf46-bc77-412b-9440-3032bfc06c57");
+      "f6bbaf46-bc77-412b-9440-3032bfc06c57");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.DATA_COMPOSITION_APPEARANCE_TEMPLATE);
@@ -209,7 +209,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalBinaryData() {
     var mdo = getMDObjectDesigner("CommonTemplates/ДвоичныеДанные.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "ДвоичныеДанные",
-            "f4ab9283-1110-4808-9cbf-40c71ebb88a2");
+      "f4ab9283-1110-4808-9cbf-40c71ebb88a2");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.BINARY_DATA);
@@ -218,7 +218,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalHtmlDocument() {
     var mdo = getMDObjectDesigner("CommonTemplates/HTML.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "HTML",
-            "5d76084a-68fa-4579-91da-17d7ffab6225");
+      "5d76084a-68fa-4579-91da-17d7ffab6225");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.HTML_DOCUMENT);
@@ -227,7 +227,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalSpreadSheetDocument() {
     var mdo = getMDObjectDesigner("CommonTemplates/ТабличныйДокумент.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "Макет",
-            "799e0ae7-f5ea-4b50-8853-e2c58ef5d9cd");
+      "799e0ae7-f5ea-4b50-8853-e2c58ef5d9cd");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.SPREADSHEET_DOCUMENT);
@@ -236,7 +236,7 @@ class MDCommonTemplateTest extends AbstractMDOTest {
   private void checkOriginalActiveDocument() {
     var mdo = getMDObjectDesigner("CommonTemplates/Active.xml");
     checkBaseField(mdo, MDCommonTemplate.class, "Active",
-            "557665fc-86f5-44e1-9801-490cea841718");
+      "557665fc-86f5-44e1-9801-490cea841718");
     checkNoChildren(mdo);
     checkNoModules(mdo);
     assertThat(((MDCommonTemplate) mdo).getTemplateType()).isEqualTo(TemplateType.ACTIVE_DOCUMENT);
