+++ src/test/java/com/github/_1c_syntax/mdclasses/metadata/ConfigurationTest.java
@@ -76,10 +76,10 @@ class ConfigurationTest {
     assertThat(configuration.isUseOrdinaryFormInManagedApplication()).isTrue();
 
     assertThat(configuration.getCopyrights()).hasSize(2)
-            .anyMatch(copyright -> copyright.getCopyrightContent().equals("Моя Программа")
-                    && copyright.getLanguage().equals("ru"))
-            .anyMatch(copyright -> copyright.getCopyrightContent().equals("My program")
-                    && copyright.getLanguage().equals("en"));
+      .anyMatch(copyright -> copyright.getCopyrightContent().equals("Моя Программа")
+        && copyright.getLanguage().equals("ru"))
+      .anyMatch(copyright -> copyright.getCopyrightContent().equals("My program")
+        && copyright.getLanguage().equals("en"));
 
     assertThat(configuration.getModulesByType()).hasSize(38);
     assertThat(configuration.getModulesBySupport()).isEmpty();
@@ -321,10 +321,10 @@ class ConfigurationTest {
     assertThat(configuration.isUseOrdinaryFormInManagedApplication()).isFalse();
 
     assertThat(configuration.getCopyrights()).hasSize(2)
-            .anyMatch(copyright -> copyright.getCopyrightContent().equals("Моя Программа")
-                    && copyright.getLanguage().equals("ru"))
-            .anyMatch(copyright -> copyright.getCopyrightContent().equals("My program")
-                    && copyright.getLanguage().equals("en"));
+      .anyMatch(copyright -> copyright.getCopyrightContent().equals("Моя Программа")
+        && copyright.getLanguage().equals("ru"))
+      .anyMatch(copyright -> copyright.getCopyrightContent().equals("My program")
+        && copyright.getLanguage().equals("en"));
 
     assertThat(configuration.getModulesByType()).hasSize(17);
     assertThat(configuration.getModulesBySupport()).isEmpty();
